feat: added function 'Shift'
