+++ index.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./style/style.css">
     <title>Virtual Keyboard</title>
 </head>
 <body>
     
-    <script src="script.js"></script>
+    <script type="module" src="./script/script.js"></script>
 </body>
 </html>
