+++ script/keys.js
@@ -1,0 +1,36 @@
+export const en = [
+    '`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=','backspace',
+    'tab', 'q', 'w', 'e', 'r', 't', 'y', 'u', 'i', 'o', 'p', '[', ']', '\\', 'delete',
+    'caps lock', 'a', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', "'", 'enter',
+    'shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/', ' ˄ ', 'shift',
+    'ctrl', 'Alt', '', 'Alt', 'ctrl', ' ˂ ', ' ˅ ', ' ˃ '
+];
+export const enShift = [
+    '~', '!', '@', '#', '$', '%', '^', '&', '*', '(', ')', '_', '+','backspace',
+    'tab', 'Q', 'W', 'E', 'R', 'T', 'Y', 'U', 'I', 'O', 'P', '{', '}', '|', 'delete',
+    'caps lock', 'A', 'S', 'D', 'F', 'G', 'H', 'J', 'K', 'L', ':', '"', 'enter',
+    'shift', 'Z', 'X', 'C', 'V', 'B', 'N', 'M', '<', '>', '?', '˄', 'shift',
+    'ctrl', 'Alt', '', 'Alt', 'ctrl', '˂', '˅', '˃'
+];
+export const ru = [
+    'ё', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=','backspace',
+    'tab', 'й', 'ц', 'у', 'к', 'е', 'н', 'г', 'ш', 'щ', 'з', 'х', 'ъ', '\\', 'delete',
+    'caps lock', 'ф', 'ы', 'в', 'а', 'п', 'р', 'о', 'л', 'д', 'ж', 'э', 'enter',
+    'shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'б', 'ю', '.', '˄', 'shift',
+    'ctrl', 'Alt', '', 'Alt', 'ctrl', '˂', '˅', '˃'
+];
+export const ruShift = [
+    'Ё', '!', '"', '№', ';', '%', ':', '?', '*', '(', ')', '_', '+','backspace',
+    'tab', 'Й', 'Ц', 'У', 'К', 'Е', 'Н', 'Г', 'Ш', 'Щ', 'З', 'Х', 'Ъ', '/', 'delete',
+    'caps lock', 'Ф', 'Ы', 'В', 'А', 'П', 'Р', 'О', 'Л', 'Д', 'Ж', 'Э', 'enter',
+    'shift', 'Я', 'Ч', 'С', 'М', 'И', 'Т', 'Ь', 'Б', 'Ю', ',', ' ˄ ', 'shift',
+    'ctrl', 'Alt', '', 'Alt', 'ctrl', ' ˂ ', ' ˅ ', ' ˃ '
+];
+
+export const codeOfKey = [
+    'Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal','Backspace',
+    'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backslash', 'Delete',
+    'CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter',
+    'ShiftLeft', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp', 'ShiftRight',
+    'ControlLeft', 'AltLeft', 'Space', 'AltRight', 'ControlRight', 'ArrowLeft', 'ArrowDown', 'ArrowRight',
+];
